server/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
youtube-transcript-api>=0.6
google-generativeai>=0.8
pydantic-settings>=2.0
httpx>=0.27
pytest>=8.0
pytest-asyncio>=0.24
